refactor(fitting): extract noise_calibration module from utils and pipeline
Move the noise-model calibration family (fit_rel_error_from_residuals,
fit_gain_from_residuals, fit_noise_model_nnls, compute_binding_slope,
compute_plate_slopes, fit_ph_slope_noise from utils.py, plus
_noise_params_converged, _plate_noise_model_from_nnls,
calibrate_noise_robust from pipeline.py) into a new
clophfit.fitting.noise_calibration module. Pure move: pipeline.py now
holds only fgls_plate_fit, importing its five helpers from the new
module; utils.py keeps outlier/curve-shape helpers and fit_trendline.
