+++ src/clophfit/fitting/pipeline.py
@@ -1,118 +1,21 @@
-"""Pipeline orchestrators for fitting multistage workflows (e.g., FGLS)."""
+"""Iterative FGLS plate fitting."""
 
 import logging
 
-import pandas as pd
-
 from clophfit.fitting.core import fit_binding_glob
-from clophfit.fitting.data_structures import (
-    Dataset,
-    FitResult,
-    NoiseModelParams,
-    PlateNoiseModel,
-)
+from clophfit.fitting.data_structures import Dataset, FitResult, PlateNoiseModel
 from clophfit.fitting.errors import InsufficientDataError
 from clophfit.fitting.model_validation import residuals_from_fit_results
 from clophfit.fitting.models import binding_1site
-from clophfit.fitting.utils import (
+from clophfit.fitting.noise_calibration import (
+    _noise_params_converged,
+    _plate_noise_model_from_nnls,
     compute_plate_slopes,
-    fit_gain_from_residuals,
     fit_noise_model_nnls,
     fit_ph_slope_noise,
-    fit_rel_error_from_residuals,
 )
 
 logger = logging.getLogger(__name__)
-
-
-def _noise_params_converged(
-    old: PlateNoiseModel,
-    new: PlateNoiseModel,
-    tol: float = 1e-3,
-) -> bool:
-    """Check whether gain and alpha converged across labels."""
-    for lbl in new:
-        if lbl not in old:
-            return False
-        for attr in ("gain", "alpha"):
-            old_val = getattr(old[lbl], attr)
-            new_val = getattr(new[lbl], attr)
-            denom = max(old_val, new_val, 1e-12)
-            if abs(new_val - old_val) / denom > tol:
-                return False
-    return True
-
-
-def _plate_noise_model_from_nnls(
-    sigma_floor: dict[str, float],
-    gains: dict[str, float],
-    alphas: dict[str, float],
-    sigma_ph: float = 0.0,
-) -> PlateNoiseModel:
-    """Build a PlateNoiseModel from NNLS output dicts."""
-    model = PlateNoiseModel()
-    for lbl in sigma_floor:
-        model[lbl] = NoiseModelParams(
-            sigma_floor=sigma_floor.get(lbl, 0.0),
-            gain=gains.get(lbl, 0.0),
-            alpha=alphas.get(lbl, 0.0),
-            sigma_ph=sigma_ph,
-        )
-    return model
-
-
-def calibrate_noise_robust(
-    residuals: pd.DataFrame,
-    sigma_floor: dict[str, float],
-    *,
-    p_threshold: float = 0.9,
-    min_keep: int = 3,
-) -> PlateNoiseModel:
-    """Calibrate a per-label noise model from outlier-screened residuals.
-
-    Drops points whose posterior outlier probability exceeds *p_threshold*
-    (from a PyMC mixture fit) and then estimates ``gain`` and ``alpha`` per
-    label with the single-term moment estimators
-    (:func:`~clophfit.fitting.utils.fit_gain_from_residuals`,
-    :func:`~clophfit.fitting.utils.fit_rel_error_from_residuals`) on the
-    retained points. Screening with the mixture's ``p_outlier`` keeps
-    outliers from inflating the estimate, while the two single-term estimators
-    avoid the gain/alpha collinearity of the joint NNLS over narrow titration
-    ranges.
-
-    Parameters
-    ----------
-    residuals : pd.DataFrame
-        Canonical residual table (e.g. ``MultiFitResult.residuals`` from a
-        mixture fit). Must have ``label``, ``raw_res``, ``yhat`` columns; a
-        ``p_outlier`` column enables screening (otherwise all points
-        are used).
-    sigma_floor : dict[str, float]
-        Known read-noise floor per label, e.g. ``tit.bg_noise``. Used as the
-        fixed floor and copied into the returned model.
-    p_threshold : float, optional
-        Posterior outlier probability above which a point is dropped.
-    min_keep : int, optional
-        Per label, if screening would retain fewer than this many points the
-        full (unscreened) set is used instead.
-
-    Returns
-    -------
-    PlateNoiseModel
-        Per-label model with ``sigma_floor`` from *sigma_floor* and calibrated
-        ``gain``/``alpha``.
-    """
-    if "p_outlier" in residuals.columns:
-        kept: list[pd.DataFrame] = []
-        for _label, group in residuals.groupby("label", observed=True):
-            inliers = group[group["p_outlier"].fillna(0.0) < p_threshold]
-            kept.append(group if len(inliers) < min_keep else inliers)
-        clean = pd.concat(kept, ignore_index=True) if kept else residuals
-    else:
-        clean = residuals
-    gains = fit_gain_from_residuals(clean, sigma_floor)
-    alphas = fit_rel_error_from_residuals(clean, sigma_floor)
-    return _plate_noise_model_from_nnls(dict(sigma_floor), gains, alphas)
 
 
 def fgls_plate_fit(  # noqa: PLR0913
